turning the myaccountpage into components to make the code more organized
